split wasmto-{js,dts}
