+++ bin/cli-node.js
@@ -1,6 +1,11 @@
-import { stdin, stdout, wasmtoJs } from 'wasmto-js';
+import { stdin, stdout, wasmtoJs, wasmtoDts } from 'wasmto-js';
 
 const reader = stdin();
 const writer = stdout();
 
-await wasmtoJs(reader, writer);
+const dts = process.argv.includes('--dts', 2);
+if (dts) {
+    await wasmtoDts(reader, writer);
+} else {
+    await wasmtoJs(reader, writer);
+}
