+++ src/index.ts
@@ -1,2 +1,3 @@
-export { wasmtoJs, wasmtoDts } from "./wasmto-js";
+export { wasmtoJs } from "./wasmto-js";
+export { wasmtoDts } from "./wasmto-dts";
 export { stdin, stdout } from "./stream";
